Fix #88

+++ kernel/include/arch/x86/pit.h
@@ -6,6 +6,8 @@
 
 #define OSCILLATOR_FREQUENCY 1193182
 
-size_t PIT_init(uint16 a_divisor);
+uint32 PIT_init(
+    uint16 a_divisor
+);
 
 #endif
